account-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-restful==0.3.9
mysql-connector-python==8.0.26
prometheus_client==0.17.1
